messenger/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.1
defusedxml==0.7.1
Django==4.1.7
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-dbbackup==4.0.2
django-timezone-field==5.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
flake8==6.0.0
flower==1.2.0
humanize==4.6.0
idna==3.4
kombu==5.2.4
mccabe==0.7.0
oauthlib==3.2.2
Pillow==9.4.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
